Fix failing tests due to changed architecture of xunit:
the assert tests where failing because the assert symbols couldn't be resolved. This was because the references added in the tests where referencing the `FactAttribute` class (which lives in `xunit.core.dll`), however, the `Assert` class is in `xunit.assert.dll`).

This adds that metadata reference in the `ConverterTestBase` class.